rename static task to assets
