+++ scripts/tasks/assets.js
